feat: add advanced nmap viewer
